Fix fetching of remote schemas.
Before when `fetchRefs` searched through a schema for "$refs", it
assumed every object in the schema's JSON was a valid subschema. This is
not the case (for instance "definitions" will be an object, but that
object's _members_ are subschemas, not the object itself.

This is now fixed so that only actual subschemas are checked for "refs".