1.3.0.0.
+ Rewrite failure messages.

We used to return a list of every failed validator, no matter how far it was buried within subschemas of the original schema.

Now we return a tree of failure messages, so that if the first schema had only three validators, then no more than three top level failures will be returned.

+ Move the code to parse each validator from JSON from `Data.Validator.Draft4` into the validators' modules themselves.

+ Switch from Prelude to Protolude.

+ Switch `Data.JsonSchema.Fetch` from lazy to strict bytestrings.