Prepare to update the test suite.
This involves a couple of fixes:

+ Validators alongside of "$ref" are now ignored.

+ Relative references are now correctly resolved against the current
document, even if the scope of the schema has been changed by `"id"`.

To support this some of hjsonschema's API has changed:

+ `ReferencedSchemas` has been removed. It used to contain both a
`URISchemaMap` and the very first schema, we now simply use plain
`URISchemaMap`s instead.

+ A new `Scope` data type has been added. It tracks the URI at which the
current document originated, the contents of the current document, and
the current base for resolution of references, taking into account
`"id"`.